Finish up the logic for obtaining the tree
