+++ priv/static/builder/deno.lock
@@ -76,6 +76,8 @@
     "https://deno.land/std@0.196.0/path/posix.ts": "8b7c67ac338714b30c816079303d0285dd24af6b284f7ad63da5b27372a2c94d",
     "https://deno.land/std@0.196.0/path/separator.ts": "0fb679739d0d1d7bf45b68dacfb4ec7563597a902edbaf3c59b50d5bcadd93b1",
     "https://deno.land/std@0.196.0/path/win32.ts": "4fca292f8d116fd6d62f243b8a61bd3d6835a9f0ede762ba5c01afe7c3c0aa12",
+    "https://deno.land/std@0.196.0/testing/asserts.ts": "b4e4b1359393aeff09e853e27901a982c685cb630df30426ed75496961931946",
+    "https://deno.land/std@0.196.0/testing/snapshot.ts": "a45165a6a2495f1ec59bf02fda4819de189b000f0009c121ea8ac9c51bae63c6",
     "https://deno.land/x/appsignal@v1.0.1/mod.ts": "361e66bffc6df8622c151040edbf2d24a8c60ef7a538c5118c132addcad898d2",
     "https://deno.land/x/ascii_table@v0.1.0/mod.ts": "bd99729ffefd9e0792787b7aafd06c5257039b31ffe3b987fe5257d3fa0c39b5",
     "https://deno.land/x/file_extension@v2.1.0/mod.ts": "a674ce4457177809fec87b64ef7109aa0fc7cd4161fb00b30e7efc3004ae48ca",
