+++ priv/static/builder/utils/vcs/utilities_test.ts
@@ -3,8 +3,10 @@ import { runInTemporaryDirectory } from "../../tests/test-helpers.ts";
 import {
   extractPlaceholderValuesFromFilePath,
   getFileFormat,
+  getFileSHA256,
 } from "./utilities.ts";
 import { assertEquals } from "https://deno.land/std@0.196.0/assert/assert_equals.ts";
+import { assertNotEquals } from "https://deno.land/std@0.196.0/assert/assert_not_equals.ts";
 
 Deno.test("getFileFormat when the format is markdown", async () => {
   await runInTemporaryDirectory(async (temporaryDirectory) => {
@@ -87,4 +89,86 @@ Deno.test("extractPlaceholdersFromFile extracts all the placeholders", () => {
   // Then
   assertEquals(got["language"], "en");
   assertEquals(got["country"], "US");
+});
+
+Deno.test("getFileSHA256 returns the same value when the content hasn't changed", async () => {
+  await runInTemporaryDirectory(async (temporaryDirectory) => {
+    // Given
+    const filePath = join(temporaryDirectory, "foo.txt");
+    await Deno.writeTextFile(filePath, "bar");
+
+    // When
+    const lhsHash = await getFileSHA256(filePath, { language: "es" });
+    const rhsHash = await getFileSHA256(filePath, { language: "es" });
+
+    // Then
+    assertEquals(lhsHash, rhsHash);
+  });
+});
+
+Deno.test("getFileSHA256 returns a different value when the content changes", async () => {
+  await runInTemporaryDirectory(async (temporaryDirectory) => {
+    // Given
+    const filePath = join(temporaryDirectory, "foo.txt");
+    await Deno.writeTextFile(filePath, "bar");
+
+    // When
+    const lhsHash = await getFileSHA256(filePath, { language: "es" });
+    await Deno.writeTextFile(filePath, "foo");
+    const rhsHash = await getFileSHA256(filePath, { language: "es" });
+
+    // Then
+    assertNotEquals(lhsHash, rhsHash);
+  });
+});
+
+Deno.test("getFileSHA256 returns a different value when the context changes", async () => {
+  await runInTemporaryDirectory(async (temporaryDirectory) => {
+    // Given
+    const filePath = join(temporaryDirectory, "foo.txt");
+    await Deno.writeTextFile(filePath, "bar");
+
+    // When
+    const lhsHash = await getFileSHA256(filePath, { language: "es" });
+    const rhsHash = await getFileSHA256(filePath, { language: "de" });
+
+    // Then
+    assertNotEquals(lhsHash, rhsHash);
+  });
+});
+
+Deno.test("getFileSHA256 returns the same value when the context doesn't change", async () => {
+  await runInTemporaryDirectory(async (temporaryDirectory) => {
+    // Given
+    const filePath = join(temporaryDirectory, "foo.txt");
+    await Deno.writeTextFile(filePath, "bar");
+
+    // When
+    const lhsHash = await getFileSHA256(filePath, { language: "es" });
+    const rhsHash = await getFileSHA256(filePath, { language: "es" });
+
+    // Then
+    assertEquals(lhsHash, rhsHash);
+  });
+});
+
+Deno.test("getFileSHA256 returns the same value regardles off the order of the context keys", async () => {
+  await runInTemporaryDirectory(async (temporaryDirectory) => {
+    // Given
+    const filePath = join(temporaryDirectory, "foo.txt");
+    await Deno.writeTextFile(filePath, "bar");
+
+    // When
+    const lhsHash = await getFileSHA256(filePath, {
+      language: "es",
+      country: "MX",
+    });
+    const rhsHash = await getFileSHA256(filePath, {
+      country: "MX",
+      language: "es",
+    });
+
+    // Then
+    assertEquals(lhsHash, rhsHash);
+  });
 });
